Move the Job running code into the model
This will allow its results to be seamlessly stored into the db
